requirements.txt
Django==2.0.7
django-jsonfield==1.0.1